requirements.txt
asgiref==3.8.1
Django==3.2.19
django-cors-headers==4.3.1
djangorestframework==3.15.1
pillow==10.3.0
sqlparse==0.5.0
django
python-dotenv
psycopg2-binary
django-cors-headers
dj-database-url
